fix all project, removed unnecessary code and cleaned up the code

+++ WebApi.DAL/Repositories/BookmarkRepository.cs
@@ -1,10 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using WebApi.DAL.Entities;
+﻿using WebApi.DAL.Entities;
 using WebApi.DAL.Interfaces;
 
 namespace WebApi.DAL.Repositories
@@ -17,7 +11,6 @@ namespace WebApi.DAL.Repositories
         {
             _db = context;
         }
-
         public List<Bookmark> BookmarksOfUser(string id)
         {
             var bookmarks = _db.Bookmarks
@@ -25,23 +18,28 @@ namespace WebApi.DAL.Repositories
                 .ToList();
             return bookmarks;
         }
-
         public void Create(Bookmark bookmark)
         {
-            var bookmarkDb = _db.Bookmarks
-                .Where(b => b.UserId == bookmark.UserId && b.PostId == bookmark.PostId)
+            var postDb = _db.Posts
+                .Where(p => p.Id == bookmark.PostId)
                 .FirstOrDefault();
-            if (bookmarkDb == null)
-            {
-                _db.Bookmarks.Add(bookmark);
-                _db.SaveChanges();
-            }
-            else
-            {
-                _db.Bookmarks.Remove(bookmarkDb);
-                _db.SaveChanges();
-            }
 
+            if (postDb != null)
+            {
+                var bookmarkDb = _db.Bookmarks
+                .Where(b => b.UserId == bookmark.UserId && b.PostId == postDb.Id)
+                .FirstOrDefault();
+                if (bookmarkDb == null)
+                {
+                    _db.Bookmarks.Add(bookmark);
+                    _db.SaveChanges();
+                }
+                else
+                {
+                    _db.Bookmarks.Remove(bookmarkDb);
+                    _db.SaveChanges();
+                }
+            }
         }
     }
 }
